# challenge_foco/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/"
        }
    ],
    "paths": {
...
        "/api/reserve": {
            "post": {
                "tags": [
                    "Reserves"
                ],
                "summary": "Get reserve by ID, POST method",
                "description": "Retrieve a reserve record using the reserve ID",
                "operationId": "2c32cd34c9babdc7684d1594715fd40c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Reserve ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "hotel_id",
                        "in": "query",
                        "description": "Hotel ID to which the reserve belongs",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "room_id",
                        "in": "query",
                        "description": "Room ID to which the reserve belongs",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserve found"
                    },
                    "404": {
                        "description": "Reserve not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/create-reserve": {
            "post": {
                "tags": [
                    "Reserves"
                ],
                "summary": "Create a new reserve",
                "description": "Create a reserve record",
                "operationId": "72598ff6898098161f04bb65ef185faa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "hotel_id",
                                    "room_id",
                                    "check_in",
                                    "check_out",
                                    "total"
                                ],
                                "properties": {
                                    "hotel_id": {
                                        "description": "ID of associated hotel",
                                        "type": "integer"
                                    },
                                    "room_id": {
                                        "description": "ID of associated room",
                                        "type": "integer"
                                    },
                                    "check_in": {
                                        "description": "Check-in date in YYYY-MM-DD format",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "check_out": {
                                        "description": "Check-out date in YYYY-MM-DD format",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "total": {
                                        "description": "Total cost of the reservation",
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reserve created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reserve created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Reserve"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Bad request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Server error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/update-reserve": {
            "post": {
                "tags": [
                    "Reserves"
                ],
                "summary": "Update a reserve",
                "description": "Update a reserve record",
                "operationId": "3bec54740b28a475de608c45de47bc3b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "description": "ID of the reserve to update",
                                        "type": "integer"
                                    },
                                    "hotel_id": {
                                        "description": "ID of associated hotel",
                                        "type": "integer"
                                    },
                                    "room_id": {
                                        "description": "ID of associated room",
                                        "type": "integer"
                                    },
                                    "check_in": {
                                        "description": "Check-in date in YYYY-MM-DD format",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "check_out": {
                                        "description": "Check-out date in YYYY-MM-DD format",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "total": {
                                        "description": "Total cost of the reservation",
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reserve updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reserve updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Reserve"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Bad request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Server error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Reserve": {
                "description": "Reserve model",
                "properties": {
                    "id": {
                        "description": "Reserve ID",
                        "type": "integer"
                    },
                    "hotel_id": {
                        "description": "ID of the associated hotel",
                        "type": "integer"
                    },
                    "room_id": {
                        "description": "ID of the associated room",
                        "type": "integer"
                    },
                    "check_in": {
                        "description": "Check-in date in YYYY-MM-DD format",
                        "type": "string",
                        "format": "date"
                    },
                    "check_out": {
                        "description": "Check-out date in YYYY-MM-DD format",
                        "type": "string",
                        "format": "date"
                    },
                    "total": {
                        "description": "Total cost of the reservation",
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}